# docs/swagger.yaml
info:
  contact: {}
paths:
  /index:
    get:
      responses:
        "200":
          description: OK
          schema:
            type: string
      tags:
      - index
  /user/createUser:
    get:
      parameters:
      - description: userName
        in: query
        name: name
        type: string
      - description: password
        in: query
        name: password
        type: string
      - description: repassword
        in: query
        name: repassword
        type: string
      responses:
        "200":
          description: code", "message"}
          schema:
            type: string
      summary: create user
      tags:
      - User
  /user/deleteUser:
    get:
      parameters:
      - description: id
        in: query
        name: id
        type: string
      responses:
        "200":
          description: code", "message"}
          schema:
            type: string
      summary: delete user
      tags:
      - User
  /user/findUserByNameAndPwd:
    post:
      parameters:
      - description: userName
        in: query
        name: name
        type: string
      - description: password
        in: query
        name: password
        type: string
      responses:
        "200":
          description: code", "message"}
          schema:
            type: string
      summary: User log in
      tags:
      - User
  /user/getUserList:
    get:
      responses:
        "200":
          description: code", "message"}
          schema:
            type: string
      summary: user list
      tags:
      - User
  /user/updateUser:
    post:
      parameters:
      - description: id
        in: formData
        name: id
        type: string
      - description: name
        in: formData
        name: name
        type: string
      - description: passWord
        in: formData
        name: password
        type: string
      - description: phone
        in: formData
        name: phone
        type: string
      - description: email
        in: formData
        name: email
        type: string
      responses:
        "200":
          description: code", "message"}
          schema:
            type: string
      summary: update user
      tags:
      - User
swagger: "2.0"